+++ src/main/java/at/porscheinformatik/seleniumcomponents/ClickableSeleniumComponent.java
@@ -38,7 +38,7 @@ public interface ClickableSeleniumComponent extends VisibleSeleniumComponent
      */
     default void click()
     {
-        LOG.info(String.format("[S] Clicking on %s", describe()));
+        LOG.interaction("Clicking on %s", describe());
 
         SeleniumAsserts.assertIsClickable(SeleniumGlobals.getShortTimeoutInSeconds(), this);
         SeleniumUtils.retryOnStale(() -> element().click());