+++ src/main/java/at/porscheinformatik/seleniumcomponents/EditableSeleniumComponent.java
@@ -61,7 +61,7 @@ public interface EditableSeleniumComponent extends ClickableSeleniumComponent
      */
     default void clear()
     {
-        LOG.info(String.format("[S] Clearing %s", describe()));
+        LOG.interaction("Clearing %s", describe());
 
         SeleniumAsserts.assertIsEditable(SeleniumGlobals.getShortTimeoutInSeconds(), this);
         SeleniumUtils.retryOnStale(() -> element().clear());
@@ -75,7 +75,7 @@ public interface EditableSeleniumComponent extends ClickableSeleniumComponent
      */
     default void sendKeys(CharSequence... keysToSend)
     {
-        LOG.info(String.format("[S] Sending \"%s\" to %s", String.join("", keysToSend), describe()));
+        LOG.interaction("Sending \"%s\" to %s", String.join("", keysToSend), describe());
 
         SeleniumAsserts.assertIsClickable(SeleniumGlobals.getShortTimeoutInSeconds(), this);
         SeleniumUtils.retryOnStale(() -> {