+++ src/main/java/at/porscheinformatik/seleniumcomponents/SelectableSeleniumComponent.java
@@ -40,7 +40,7 @@ public interface SelectableSeleniumComponent extends SeleniumComponent
     {
         if (!isSelected())
         {
-            LOG.info(String.format("[S] Selecting %s", describe()));
+            LOG.interaction("Selecting %s", describe());
 
             element().click();
         }
@@ -53,8 +53,8 @@ public interface SelectableSeleniumComponent extends SeleniumComponent
     {
         if (isSelected())
         {
-            LOG.info(String.format("[S] Deselecting %s", describe()));
-            
+            LOG.interaction("Deselecting %s", describe());
+
             element().click();
         }
     }