+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/SelectComponent.java
@@ -47,7 +47,7 @@ public class SelectComponent extends AbstractSeleniumComponent
 
     public void selectByLabel(String label)
     {
-        LOG.info(String.format("[S] Selecting \"%s\" of %s by label", label, describe()));
+        LOG.interaction("Selecting \"%s\" of %s by label", label, describe());
 
         WebElement element = element();
 
@@ -58,7 +58,7 @@ public class SelectComponent extends AbstractSeleniumComponent
 
     public void selectByValue(String value)
     {
-        LOG.info(String.format("[S] Selecting \"%s\" of %s by value", value, describe()));
+        LOG.interaction("Selecting \"%s\" of %s by value", value, describe());
 
         OptionComponent option = optionsFactory.find($ -> Objects.equals(value, $.getValue()));